[Rename] : Community -> ccc 폴더 이름 변경
