+++ client/src/pages/Community/ReiviewDetail.tsx
@@ -11,14 +11,14 @@ import Swal from 'sweetalert2';
 
 import { Viewer } from '@toast-ui/react-editor';
 
-import Answers from '../../Components/Community/Answers';
-import MapApi from '../../Components/Community/MapApi';
-import ReviewCarousel from '../../Components/Community/ReviewCarousel';
+import Answers from '../../Components/ccc/Answers';
+import MapApi from '../../Components/ccc/MapApi';
+import ReviewCarousel from '../../Components/ccc/ReviewCarousel';
 import { Api } from '../../apis/customAPI';
 import useAxios from '../../hooks/useAxios';
 import { Iuser } from '../../reducers/userInfoReducer';
 import { RootState } from '../../store/Store';
-import TopBar from '../../Components/Community/TopBar';
+import TopBar from '../../Components/ccc/TopBar';
 
 const TopBarContainer = styled.div`
 	margin-top: -80px;
